refactor(skill-registry): add debug logging and move default_sources to ops.rs
- Add tracing::debug logging to list_sources, add_source, remove_source,
  and search_catalog per project coding guidelines.
- Move default_sources() from types.rs to ops.rs since types.rs should
  only contain Serde domain types.
